+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] – 2020-09-19
+
 ### Added
 
 - Add support for additional binary hardlinks (see [GH-12]).
@@ -62,4 +64,5 @@ Initial prerelease.
 [0.0.3]: https://github.com/lunaryorn/homebins/compare/v0.0.2...v0.0.3
 [0.0.4]: https://github.com/lunaryorn/homebins/compare/v0.0.3...v0.0.4
 [0.0.5]: https://github.com/lunaryorn/homebins/compare/v0.0.4...v0.0.5
-[Unreleased]: https://github.com/lunaryorn/homebins/compare/v0.0.5...HEAD
+[0.1.0]: https://github.com/lunaryorn/homebins/compare/v0.0.5...v0.1.0
+[Unreleased]: https://github.com/lunaryorn/homebins/compare/v0.1.0...HEAD